# apps/api/src/modules/payments.rs
use axum::extract::{Path, Query, State};
use axum::http::StatusCode;
use axum::response::Response;
use axum::{Json, Router};
use chrono::{DateTime, Utc};
use serde::{Deserialize, Serialize};
use serde_json::json;
use sqlx::{PgPool, QueryBuilder, Row};
use uuid::Uuid;

use crate::errors::{map_database_error, ApiError};
use crate::extractors::ApiJson;
use crate::helpers::resolve_commune_filter;
use crate::modules::audit;
use crate::modules::auth::AuthUser;
use crate::modules::pvs::{load_pv, record_status_change_tx};
use crate::modules::rbac::Role;
use crate::pagination::{Paginated, Pagination, PaginationQuery};
use crate::sequences::{next_document_sequence, SEQUENCE_RECEIPT};
use crate::state::AppState;

// ─────────────────────────────────────────────────────────────────────────────
// Router
// ─────────────────────────────────────────────────────────────────────────────

pub fn router() -> Router<AppState> {
    Router::new()
        .route("/payments", axum::routing::get(list_payments))
        .route("/payments/pending", axum::routing::get(list_pending))
        .route("/payments/summary", axum::routing::get(caisse_summary))
        .route(
            "/payments/{pv_id}/validate",
            axum::routing::post(validate_payment),
        )
        .route(
            "/payments/{id}/receipt",
            axum::routing::get(get_receipt_pdf),
        )
}

// ─────────────────────────────────────────────────────────────────────────────
// Types
// ─────────────────────────────────────────────────────────────────────────────

#[derive(Debug, Serialize, Clone)]
pub struct PaymentResponse {
    pub id: Uuid,
    pub pv_id: Uuid,
    pub commune_id: Uuid,
    pub amount_due: f64,
    pub amount_penalty: f64,
    pub amount_total: f64,
    pub amount_paid: f64,
    pub amount_due_fcfa: Option<i64>,
    pub amount_penalty_fcfa: i64,
    pub amount_total_fcfa: Option<i64>,
    pub amount_paid_fcfa: Option<i64>,
    pub receiver_user_id: Uuid,
    pub paid_at: Option<DateTime<Utc>>,
    pub status: String,
    pub receipt_number: Option<String>,
    pub created_at: DateTime<Utc>,
}

#[derive(Debug, Serialize)]
pub struct PendingPvResponse {
    pub pv_id: Uuid,
    pub pv_number: String,
    pub commune_id: Uuid,
    pub amount_due: f64,
    pub amount_penalty: f64,
    pub amount_total: f64,
    pub amount_due_fcfa: Option<i64>,
    pub amount_penalty_fcfa: i64,
    pub amount_total_fcfa: Option<i64>,
    /// Alias explicite de `amount_due_fcfa` — la « base » de l'amende, que le client
    /// veut pouvoir lire séparément de la pénalité.
    pub amount_base_fcfa: i64,
    /// Retard **dérivé des dates**, et non du statut : un PV échu dont l'infraction a un
    /// taux nul est en retard sans porter de pénalité, cas que l'ancienne heuristique
    /// côté front (`amount_penalty_fcfa > 0`) manquait.
    pub is_late: bool,
    pub due_date: DateTime<Utc>,
    pub created_at: DateTime<Utc>,
}

#[derive(Debug, Deserialize)]
pub struct PaymentFilterQuery {
    page: Option<i64>,
    page_size: Option<i64>,
    commune_id: Option<Uuid>,
    status: Option<String>,
}

#[derive(Debug, Deserialize)]
pub struct ValidatePaymentRequest {
    pub amount_paid_fcfa: Option<i64>,
    pub amount_paid: Option<f64>,
}

/// Montants d'un PV tels que servis par la vue `pv_amounts_due`.
/// `list_pending` lit la vue directement (projection complète) ; cette structure ne porte
/// que ce dont l'encaissement a besoin.
struct PaymentComputation {
    amount_due_fcfa: i64,
    amount_penalty_fcfa: i64,
    amount_total_fcfa: i64,
}

/// Agrégats de la page « Caisse & paiements ».
///
/// Auparavant calculés dans le navigateur sur les 50 dernières lignes et **sans filtrer
/// `payments.status`** : un paiement annulé entrait dans la recette du jour et le total
/// divergeait de celui du tableau de bord. Tout est désormais agrégé en SQL.
#[derive(Debug, Serialize)]
pub struct CaisseSummaryResponse {
    pub day: chrono::NaiveDate,
    pub timezone: String,
    pub receipts_today: i64,
    pub collected_today_fcfa: i64,
    pub collected_today_base_fcfa: i64,
    pub collected_today_penalty_fcfa: i64,
    pub pending_count: i64,
    pub pending_late_count: i64,
    pub pending_base_fcfa: i64,
    pub pending_penalty_fcfa: i64,
    pub pending_total_fcfa: i64,
    pub commune_id: Option<Uuid>,
}

#[derive(Debug, Deserialize)]
pub struct CaisseSummaryQuery {
    commune_id: Option<Uuid>,
    date: Option<chrono::NaiveDate>,
}

// ─────────────────────────────────────────────────────────────────────────────
// Handlers
// ─────────────────────────────────────────────────────────────────────────────

async fn list_payments(
    State(state): State<AppState>,
    auth_user: AuthUser,
    Query(query): Query<PaymentFilterQuery>,
) -> Result<Json<Paginated<PaymentResponse>>, ApiError> {
    auth_user.require_any_role(&[
        Role::SuperAdmin,
        Role::AdminCommune,
        Role::Superviseur,
        Role::Receveur,
    ])?;

    let pagination = Pagination::from_query(PaginationQuery {
        page: query.page,
        page_size: query.page_size,
    })?;

    let commune_filter = resolve_commune_filter(&auth_user, query.commune_id)?;

    let mut count_qb: QueryBuilder<sqlx::Postgres> =
        QueryBuilder::new("SELECT COUNT(*) AS total FROM payments WHERE 1=1");
    if let Some(id) = commune_filter {
        count_qb.push(" AND commune_id = ").push_bind(id);
    }
    if let Some(ref s) = query.status {
        count_qb.push(" AND status = ").push_bind(s.clone());
    }
    let total: i64 = count_qb.build().fetch_one(&state.db).await?.get("total");

    let mut qb: QueryBuilder<sqlx::Postgres> = QueryBuilder::new(
        r#"
        SELECT
            id, pv_id, commune_id,
            amount_due::DOUBLE PRECISION AS amount_due,
            amount_penalty::DOUBLE PRECISION AS amount_penalty,
            amount_total::DOUBLE PRECISION AS amount_total,
            amount_paid::DOUBLE PRECISION AS amount_paid,
            amount_due_fcfa, amount_penalty_fcfa, amount_total_fcfa,
            amount_paid_fcfa, receiver_user_id, paid_at, status,
            receipt_number, created_at
        FROM payments
        WHERE 1=1
        "#,
    );
    if let Some(id) = commune_filter {
        qb.push(" AND commune_id = ").push_bind(id);
    }
    if let Some(ref s) = query.status {
        qb.push(" AND status = ").push_bind(s.clone());
    }
    qb.push(" ORDER BY created_at DESC LIMIT ")
        .push_bind(pagination.limit)
        .push(" OFFSET ")
        .push_bind(pagination.offset);

    let rows = qb.build().fetch_all(&state.db).await?;
    let items = rows.into_iter().map(row_to_payment).collect();
    Ok(Json(Paginated::new(items, &pagination, total)))
}

async fn list_pending(
    State(state): State<AppState>,
    auth_user: AuthUser,
    Query(query): Query<PaymentFilterQuery>,
) -> Result<Json<Paginated<PendingPvResponse>>, ApiError> {
    auth_user.require_any_role(&[Role::SuperAdmin, Role::AdminCommune, Role::Receveur])?;

    let pagination = Pagination::from_query(PaginationQuery {
        page: query.page,
        page_size: query.page_size,
    })?;

    let commune_filter = resolve_commune_filter(&auth_user, query.commune_id)?;

    // Comptage et page lisent le MÊME prédicat `is_pending` de la vue : `total` et
    // `items` ne peuvent donc plus diverger. La boucle par PV (une requête chacun) a
    // disparu — les montants sont calculés en une passe par la vue.
    let mut count_qb: QueryBuilder<sqlx::Postgres> =
        QueryBuilder::new("SELECT COUNT(*) AS total FROM pv_amounts_due WHERE is_pending");
    if let Some(id) = commune_filter {
        count_qb.push(" AND commune_id = ").push_bind(id);
    }
    let total: i64 = count_qb.build().fetch_one(&state.db).await?.get("total");

    let mut qb: QueryBuilder<sqlx::Postgres> = QueryBuilder::new(
        r#"
        SELECT
            pv_id, pv_number, commune_id,
            amount_base_fcfa, amount_penalty_fcfa, amount_total_fcfa,
            due_date, is_late, created_at
        FROM pv_amounts_due
        WHERE is_pending
        "#,
    );

    if let Some(id) = commune_filter {
        qb.push(" AND commune_id = ").push_bind(id);
    }
    qb.push(" ORDER BY created_at ASC LIMIT ")
        .push_bind(pagination.limit)
        .push(" OFFSET ")
        .push_bind(pagination.offset);

    let rows = qb.build().fetch_all(&state.db).await?;
    let items = rows
        .into_iter()
        .map(|row| {
            let base: i64 = row.get("amount_base_fcfa");
            let penalty: i64 = row.get("amount_penalty_fcfa");
            let total_fcfa: i64 = row.get("amount_total_fcfa");
            PendingPvResponse {
                pv_id: row.get("pv_id"),
                pv_number: row.get("pv_number"),
                commune_id: row.get("commune_id"),
                amount_due: base as f64,
                amount_penalty: penalty as f64,
                amount_total: total_fcfa as f64,
                amount_due_fcfa: Some(base),
                amount_penalty_fcfa: penalty,
                amount_total_fcfa: Some(total_fcfa),
                amount_base_fcfa: base,
                is_late: row.get("is_late"),
                due_date: row.get("due_date"),
                created_at: row.get("created_at"),
            }
        })
        .collect();

    Ok(Json(Paginated::new(items, &pagination, total)))
}

/// Agrégats serveur de la caisse : recette du jour (statut `PAYE` uniquement) et
/// encours base/pénalité/total, pour que la page Caisse et le tableau de bord comptent
/// exactement la même chose.
async fn caisse_summary(
    State(state): State<AppState>,
    auth_user: AuthUser,
    Query(query): Query<CaisseSummaryQuery>,
) -> Result<Json<CaisseSummaryResponse>, ApiError> {
    auth_user.require_any_role(&[
        Role::SuperAdmin,
        Role::AdminCommune,
        Role::Superviseur,
        Role::Receveur,
    ])?;

    let commune_filter = resolve_commune_filter(&auth_user, query.commune_id)?;
    let timezone = state.config.app_timezone.clone();

    // Le jour courant est déterminé côté SQL dans le fuseau de la commune : `paid_at`
    // est en UTC (WAT = UTC+1) et le front utilisait l'heure locale du poste, si bien
    // qu'un encaissement à 00h30 était compté la veille.
    let day = match query.date {
        Some(date) => date,
        None => sqlx::query_scalar::<_, chrono::NaiveDate>("SELECT (now() AT TIME ZONE $1)::date")
            .bind(&timezone)
            .fetch_one(&state.db)
            .await
            .map_err(|_| {
                ApiError::bad_request(format!("Fuseau horaire invalide: {timezone}"))
            })?,
    };

    let today_row = sqlx::query(
        r#"
        SELECT
            COUNT(*)                                      AS receipts_today,
            COALESCE(SUM(amount_paid_fcfa), 0)::BIGINT    AS collected_today,
            COALESCE(SUM(amount_due_fcfa), 0)::BIGINT     AS collected_today_base,
            COALESCE(SUM(amount_penalty_fcfa), 0)::BIGINT AS collected_today_penalty
        FROM payments
        WHERE status = 'PAYE'
          AND paid_at >= ($3::date::timestamp AT TIME ZONE $2)
          AND paid_at <  (($3::date + 1)::timestamp AT TIME ZONE $2)
          AND ($1::uuid IS NULL OR commune_id = $1)
        "#,
    )
    .bind(commune_filter)
    .bind(&timezone)
    .bind(day)
    .fetch_one(&state.db)
    .await?;

    let pending_row = sqlx::query(
        r#"
        SELECT
            COUNT(*)                                      AS pending_count,
            COUNT(*) FILTER (WHERE is_late)               AS pending_late_count,
            COALESCE(SUM(amount_base_fcfa), 0)::BIGINT    AS pending_base,
            COALESCE(SUM(amount_penalty_fcfa), 0)::BIGINT AS pending_penalty,
            COALESCE(SUM(amount_total_fcfa), 0)::BIGINT   AS pending_total
        FROM pv_amounts_due
        WHERE is_pending
          AND ($1::uuid IS NULL OR commune_id = $1)
        "#,
    )
    .bind(commune_filter)
    .fetch_one(&state.db)
    .await?;

    Ok(Json(CaisseSummaryResponse {
        day,
        timezone,
        receipts_today: today_row.get("receipts_today"),
        collected_today_fcfa: today_row.get("collected_today"),
        collected_today_base_fcfa: today_row.get("collected_today_base"),
        collected_today_penalty_fcfa: today_row.get("collected_today_penalty"),
        pending_count: pending_row.get("pending_count"),
        pending_late_count: pending_row.get("pending_late_count"),
        pending_base_fcfa: pending_row.get("pending_base"),
        pending_penalty_fcfa: pending_row.get("pending_penalty"),
        pending_total_fcfa: pending_row.get("pending_total"),
        commune_id: commune_filter,
    }))
}

async fn validate_payment(
    State(state): State<AppState>,
    auth_user: AuthUser,
    Path(pv_id): Path<Uuid>,
    ApiJson(payload): ApiJson<ValidatePaymentRequest>,
) -> Result<(StatusCode, Json<PaymentResponse>), ApiError> {
    auth_user.require_any_role(&[Role::Receveur])?;

    let commune_id = auth_user
        .commune_id
        .ok_or_else(|| ApiError::forbidden("Receveur non rattache a une commune"))?;

    let amount_paid_fcfa = match payload.amount_paid_fcfa {
        Some(amount) => amount,
        None => payload
            .amount_paid
            .map(|amount| amount.round() as i64)
            .ok_or_else(|| ApiError::bad_request("amount_paid_fcfa est requis"))?,
    };
    if amount_paid_fcfa <= 0 {
        return Err(ApiError::bad_request(
            "Le montant encaisse doit etre positif",
        ));
    }
    let amount_paid = payload.amount_paid.unwrap_or(amount_paid_fcfa as f64);

    let mut tx = state.db.begin().await?;

    let pv_row = sqlx::query(
        r#"
        SELECT
            id, commune_id, status, pv_number, amount_initial_fcfa, created_at
        FROM pvs
        WHERE id = $1 AND deleted_at IS NULL
        FOR UPDATE
        "#,
    )
    .bind(pv_id)
    .fetch_optional(&mut *tx)
    .await?
    .ok_or_else(|| ApiError::not_found("PV introuvable"))?;

    let pv_commune_id: Uuid = pv_row.get("commune_id");
    if pv_commune_id != commune_id {
        return Err(ApiError::forbidden(
            "Vous ne pouvez valider que les PV de votre commune",
        ));
    }

    let pv_status: String = pv_row.get("status");
    if pv_status != "EN_ATTENTE_PAIEMENT" && pv_status != "EN_RETARD" {
        return Err(ApiError::conflict(format!(
            "Ce PV n'est pas en attente de paiement (statut: {})",
            pv_status
        )));
    }

    let existing_payment: Option<Uuid> = sqlx::query_scalar(
        "SELECT id FROM payments WHERE pv_id = $1 AND status = 'PAYE' FOR UPDATE",
    )
    .bind(pv_id)
    .fetch_optional(&mut *tx)
    .await?;

    if existing_payment.is_some() {
        return Err(ApiError::conflict("Ce PV a deja un paiement valide"));
    }

    let computed = payment_computation_tx(&mut tx, pv_id).await?;
    let amount_due_fcfa = computed.amount_due_fcfa;
    let amount_penalty_fcfa = computed.amount_penalty_fcfa;
    let amount_total_fcfa = computed.amount_total_fcfa;
    let amount_due = amount_due_fcfa as f64;
    let amount_penalty = amount_penalty_fcfa as f64;
    let amount_total = amount_total_fcfa as f64;

    // Montant EXACT exigé. Il n'existe pas de paiement partiel (`payments.pv_id` est
    // UNIQUE) ; accepter un versement supérieur l'enregistrait intégralement et gonflait
    // la recette sans contrepartie comptable. Avec l'égalité stricte,
    // « encaissé = somme des totaux dus » devient une invariante.
    if amount_paid_fcfa < amount_total_fcfa {
        return Err(ApiError::bad_request(format!(
            "Montant insuffisant: {} FCFA requis (dont {} FCFA de penalites), {} FCFA recu",
            amount_total_fcfa, amount_penalty_fcfa, amount_paid_fcfa
        )));
    }
    if amount_paid_fcfa > amount_total_fcfa {
        return Err(ApiError::bad_request(format!(
            "Montant excedentaire: {} FCFA requis (dont {} FCFA de penalites), {} FCFA recu",
            amount_total_fcfa, amount_penalty_fcfa, amount_paid_fcfa
        )));
    }

    let receipt_number = generate_receipt_number(&mut tx, commune_id).await?;
    let payment_id = Uuid::new_v4();

    sqlx::query(
        r#"
        INSERT INTO payments (
            id, pv_id, commune_id, amount_due, amount_penalty, amount_total,
            amount_paid, amount_due_fcfa, amount_penalty_fcfa, amount_total_fcfa,
            amount_paid_fcfa, receiver_user_id, paid_at, status, receipt_number
        ) VALUES (
            $1, $2, $3, $4, $5, $6,
            $7, $8, $9, $10,
            $11, $12, now(), 'PAYE', $13
        )
        "#,
    )
    .bind(payment_id)
    .bind(pv_id)
    .bind(commune_id)
    .bind(amount_due)
    .bind(amount_penalty)
    .bind(amount_total)
    .bind(amount_paid)
    .bind(amount_due_fcfa)
    .bind(amount_penalty_fcfa)
    .bind(amount_total_fcfa)
    .bind(amount_paid_fcfa)
    .bind(auth_user.id)
    .bind(&receipt_number)
    .execute(&mut *tx)
    .await
    .map_err(map_database_error)?;

    sqlx::query("UPDATE pvs SET status = 'PAYE', updated_at = now() WHERE id = $1")
        .bind(pv_id)
        .execute(&mut *tx)
        .await?;

    record_status_change_tx(
        &mut tx,
        pv_id,
        Some(&pv_status),
        "PAYE",
        auth_user.id,
        Some("Paiement valide"),
    )
    .await;

    audit::record_for_commune_tx(
        &mut tx,
        Some(commune_id),
        Some(auth_user.id),
        "PAYMENT_VALIDATED",
        "payments",
        Some(payment_id),
        None,
        Some(json!({
            "pv_id": pv_id,
            "pv_number": pv_row.get::<String, _>("pv_number"),
            "amount_due_fcfa": amount_due_fcfa,
            "amount_penalty_fcfa": amount_penalty_fcfa,
            "amount_total_fcfa": amount_total_fcfa,
            "amount_paid_fcfa": amount_paid_fcfa,
            "receipt_number": receipt_number
        })),
        auth_user.ip_address.clone(),
        auth_user.user_agent.clone(),
    )
    .await;

    tx.commit().await?;

    Ok((
        StatusCode::CREATED,
        Json(load_payment(&state.db, payment_id).await?),
    ))
}
async fn get_receipt_pdf(
    State(state): State<AppState>,
    auth_user: AuthUser,
    Path(id): Path<Uuid>,
) -> Result<Response, ApiError> {
    auth_user.require_any_role(&[Role::SuperAdmin, Role::AdminCommune, Role::Receveur])?;

    let payment = load_payment(&state.db, id).await?;
    auth_user.require_commune_access(payment.commune_id)?;

    let pv = load_pv(&state.db, payment.pv_id).await?;
    let pdf_bytes = crate::modules::pdf::generate_receipt_pdf(&state.db, &payment, &pv).await?;

    use axum::http::header;
    use axum::response::IntoResponse;
    let filename = format!(
        "Recu-{}.pdf",
        payment.receipt_number.as_deref().unwrap_or(&id.to_string())
    );
    Ok((
        [
            (header::CONTENT_TYPE, "application/pdf".to_string()),
            (
                header::CONTENT_DISPOSITION,
                format!("attachment; filename=\"{filename}\""),
            ),
        ],
        pdf_bytes,
    )
        .into_response())
}

// ─────────────────────────────────────────────────────────────────────────────
// Helpers
// ─────────────────────────────────────────────────────────────────────────────

pub async fn load_payment(pool: &PgPool, id: Uuid) -> Result<PaymentResponse, ApiError> {
    let row = sqlx::query(
        r#"
        SELECT
            id, pv_id, commune_id,
            amount_due::DOUBLE PRECISION AS amount_due,
            amount_penalty::DOUBLE PRECISION AS amount_penalty,
            amount_total::DOUBLE PRECISION AS amount_total,
            amount_paid::DOUBLE PRECISION AS amount_paid,
            amount_due_fcfa, amount_penalty_fcfa, amount_total_fcfa,
            amount_paid_fcfa, receiver_user_id, paid_at, status,
            receipt_number, created_at
        FROM payments
        WHERE id = $1
        "#,
    )
        .bind(id)
        .fetch_optional(pool)
        .await?
        .ok_or_else(|| ApiError::not_found("Paiement introuvable"))?;
    Ok(row_to_payment(row))
}

fn row_to_payment(row: sqlx::postgres::PgRow) -> PaymentResponse {
    PaymentResponse {
        id: row.get("id"),
        pv_id: row.get("pv_id"),
        commune_id: row.get("commune_id"),
        amount_due: row.get("amount_due"),
        amount_penalty: row.get("amount_penalty"),
        amount_total: row.get("amount_total"),
        amount_paid: row.get("amount_paid"),
        amount_due_fcfa: row.get("amount_due_fcfa"),
        amount_penalty_fcfa: row.get("amount_penalty_fcfa"),
        amount_total_fcfa: row.get("amount_total_fcfa"),
        amount_paid_fcfa: row.get("amount_paid_fcfa"),
        receiver_user_id: row.get("receiver_user_id"),
        paid_at: row.get("paid_at"),
        status: row.get("status"),
        receipt_number: row.get("receipt_number"),
        created_at: row.get("created_at"),
    }
}

/// Lecture des montants dus depuis la vue `pv_amounts_due` (migration 26), **seule
/// autorité** sur la règle base / pénalité / total.
///
/// La règle était auparavant dupliquée ici en Rust : le tableau de bord sommait de son
/// côté `pvs.amount_initial_fcfa` et ne pouvait donc jamais refléter une pénalité. En
/// lisant tous la vue, le montant affiché au receveur, celui exigé à la validation et
/// l'agrégat du tableau de bord ne peuvent plus diverger.
const PV_AMOUNTS_SQL: &str = r#"
    SELECT amount_base_fcfa, amount_penalty_fcfa, amount_total_fcfa
    FROM pv_amounts_due
    WHERE pv_id = $1
"#;

fn computation_from_row(row: &sqlx::postgres::PgRow) -> PaymentComputation {
    PaymentComputation {
        amount_due_fcfa: row.get("amount_base_fcfa"),
        amount_penalty_fcfa: row.get("amount_penalty_fcfa"),
        amount_total_fcfa: row.get("amount_total_fcfa"),
    }
}

async fn payment_computation_tx(
    tx: &mut sqlx::Transaction<'_, sqlx::Postgres>,
    pv_id: Uuid,
) -> Result<PaymentComputation, ApiError> {
    // `now()` vaut `transaction_timestamp()` : le montant reste stable entre le verrou
    // `FOR UPDATE` posé plus haut et l'INSERT du paiement.
    let row = sqlx::query(PV_AMOUNTS_SQL)
        .bind(pv_id)
        .fetch_optional(&mut **tx)
        .await?
        .ok_or_else(|| ApiError::not_found("PV introuvable"))?;
    Ok(computation_from_row(&row))
}

async fn generate_receipt_number(
    tx: &mut sqlx::Transaction<'_, sqlx::Postgres>,
    commune_id: Uuid,
) -> Result<String, ApiError> {
    let commune_code: String = sqlx::query_scalar("SELECT code FROM communes WHERE id = $1")
        .bind(commune_id)
        .fetch_one(&mut **tx)
        .await?;
    // Le compteur RECEIPT peut être en retard sur des reçus déjà en base
    // (données semées / reprises) : sauter les numéros pris plutôt que de
    // laisser l'INSERT violer l'unicité — le rollback annulerait aussi
    // l'incrément et chaque nouvel essai régénérerait le même numéro.
    for _ in 0..100 {
        let (year, seq) = next_document_sequence(tx, commune_id, SEQUENCE_RECEIPT).await?;
        let candidate = format!("REC-{}-{}-{:06}", commune_code.to_uppercase(), year, seq);
        let exists: bool =
            sqlx::query_scalar("SELECT EXISTS(SELECT 1 FROM payments WHERE receipt_number = $1)")
                .bind(&candidate)
                .fetch_one(&mut **tx)
                .await?;
        if !exists {
            return Ok(candidate);
        }
    }
    Err(ApiError::internal(
        "Impossible de generer un numero de recu unique",
    ))
}