# packages/api-contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "APMTRACK API",
    "version": "0.1.0",
    "description": "Contrat OpenAPI backend APMTRACK: sante, auth, RBAC, communes, utilisateurs, agents, zones, referentiel, PV, paiements, signalements, patrouilles, dashboard, audit logs et exports."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Developpement local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/payments": {
      "get": {
        "summary": "Liste les paiements",
        "responses": {
          "200": {
            "description": "Succes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedGeneric"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/payments/pending": {
      "get": {
        "summary": "Liste les PV en attente de paiement",
        "responses": {
          "200": {
            "description": "Succes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedGeneric"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/payments/{pv_id}/validate": {
      "post": {
        "summary": "Valide un paiement",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidatePaymentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Succes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payment"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/v1/payments/{id}/receipt": {
      "get": {
        "summary": "Genere le recu PDF",
        "responses": {
          "200": {
            "description": "Document PDF",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/v1/payments/summary": {
      "get": {
        "summary": "Agregats de caisse (recette du jour et encours base/penalite/total)",
        "parameters": [
          {
            "name": "commune_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Journee comptable a agreger, dans le fuseau APP_TIMEZONE. Defaut: aujourd'hui."
          }
        ],
        "responses": {
          "200": {
            "description": "Agregats de caisse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaisseSummary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Authentification invalide",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Role ou commune non autorise",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "Contrainte d'unicite violee",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Requete invalide",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "details"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            }
          }
        }
      },
...
      "GenericObject": {
        "type": "object",
        "additionalProperties": true
      },
      "FcfaAmount": {
        "type": "integer",
        "format": "int64",
        "minimum": 0,
        "description": "Montant exact en FCFA"
      },
      "PaginatedGeneric": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenericObject"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          }
        }
      },
...
      "Payment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "pv_id": {
            "type": "string",
            "format": "uuid"
          },
          "commune_id": {
            "type": "string",
            "format": "uuid"
          },
          "amount_due": {
            "type": "number",
            "deprecated": true
          },
          "amount_due_fcfa": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FcfaAmount"
              },
              {
                "type": "null"
              }
            ]
          },
          "amount_penalty": {
            "type": "number",
            "deprecated": true
          },
          "amount_penalty_fcfa": {
            "$ref": "#/components/schemas/FcfaAmount"
          },
          "amount_total": {
            "type": "number",
            "deprecated": true
          },
          "amount_total_fcfa": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FcfaAmount"
              },
              {
                "type": "null"
              }
            ]
          },
          "amount_paid": {
            "type": "number",
            "deprecated": true
          },
          "amount_paid_fcfa": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FcfaAmount"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "type": "string"
          },
          "receipt_number": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ValidatePaymentRequest": {
        "type": "object",
        "required": [
          "amount_paid_fcfa"
        ],
        "properties": {
          "amount_paid_fcfa": {
            "$ref": "#/components/schemas/FcfaAmount"
          },
          "amount_paid": {
            "type": "number",
            "deprecated": true
          }
        }
      },
...
      "CaisseSummary": {
        "type": "object",
        "properties": {
          "day": {
            "type": "string",
            "format": "date"
          },
          "timezone": {
            "type": "string"
          },
          "receipts_today": {
            "type": "integer"
          },
          "collected_today_fcfa": {
            "type": "integer",
            "format": "int64"
          },
          "collected_today_base_fcfa": {
            "type": "integer",
            "format": "int64"
          },
          "collected_today_penalty_fcfa": {
            "type": "integer",
            "format": "int64"
          },
          "pending_count": {
            "type": "integer"
          },
          "pending_late_count": {
            "type": "integer"
          },
          "pending_base_fcfa": {
            "type": "integer",
            "format": "int64"
          },
          "pending_penalty_fcfa": {
            "type": "integer",
            "format": "int64"
          },
          "pending_total_fcfa": {
            "type": "integer",
            "format": "int64"
          },
          "commune_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
...
    }
  }
}